refactor :separate types from index.ts
