+++ .prettierrc.json
@@ -5,5 +5,6 @@
   "singleQuote": true,
   "tabWidth": 2,
   "trailingComma": "es5",
-  "useTabs": false
+  "useTabs": false,
+  "endOfLine": "auto"
 }
