+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@matin_mortazavi/react-otp-input",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "A fully customizable, one-time password input component for the web built with React",
   "main": "lib/index.js",
   "module": "lib/index.esm.js",
